[#234] 제휴 REST API 단위 테스트 코드 설계
- EATSSUNetwork 모듈로 Foundation 코드 일부 이관

+++ EATSSUNetwork/EATSSUNetwork/Sources/DTO/PartnershipResponse.swift
@@ -11,6 +11,6 @@ struct PartnershipResponse: Codable {
     let id: Int
     let partnershipType, storeName, description, startDate: String
     let endDate, restaurantType: String
-    let longitude, latitude: Int
+    let longitude, latitude: Double
     let collegeNames, departmentNames: [String]
 }
